+++ contracts/finality/src/exec/mod.rs
@@ -1,3 +1,2 @@
-pub mod admin;
 pub mod finality;
 pub mod public_randomness;