+++ contracts/finality/src/msg.rs
@@ -12,7 +12,6 @@ use {
 pub struct InstantiateMsg {
     pub admin: String,
     pub bsn_id: String,
-    pub is_enabled: bool,
     pub min_pub_rand: u64,
 }
 
@@ -38,8 +37,6 @@ pub enum QueryMsg {
     Admin {},
     #[returns(Config)]
     Config {},
-    #[returns(bool)]
-    IsEnabled {},
 }
 
 // Note: Adapted from packages/apis/src/btc_staking_api.rs / packages/apis/src/finality_api.rs
@@ -100,13 +97,6 @@ pub enum ExecuteMsg {
     },
 
     // SHOULD: Administrative messages
-    /// Set enabled status of the finality contract.
-    ///
-    /// This message can be called by the admin only.
-    /// If disabled, the finality contract and the BTC staking finality will not be used
-    /// by the rollup. Note this should be implemented in the rollup's finality gadget daemon
-    /// program and is not enforced by the contract itself.
-    SetEnabled { enabled: bool },
     /// Update the admin address.
     ///
     /// This message can be called by the admin only.