+++ contracts/finality/src/state/config.rs
@@ -6,8 +6,6 @@ use cw_storage_plus::Item;
 
 pub const ADMIN: Admin = Admin::new("admin");
 pub const CONFIG: Item<Config> = Item::new("config");
-// if the finality gadget is disabled, it will always return true for the is finalized query
-pub const IS_ENABLED: Item<bool> = Item::new("is_enabled");
 
 /// Config are OP finality gadget's configuration
 #[cw_serde]
@@ -18,10 +16,6 @@ pub struct Config {
 
 pub fn get_config(deps: Deps<BabylonQuery>) -> StdResult<Config> {
     CONFIG.load(deps.storage)
-}
-
-pub fn get_is_enabled(deps: Deps<BabylonQuery>) -> StdResult<bool> {
-    IS_ENABLED.load(deps.storage)
 }
 
 pub fn get_admin(deps: Deps<BabylonQuery>) -> StdResult<AdminResponse> {