+++ contracts/finality/tests/integration.rs
@@ -1,12 +1,12 @@
 use cosmwasm_std::{from_json, Coin, ContractResult, Response};
 use cosmwasm_vm::testing::{
-    execute, instantiate, mock_env, mock_info, mock_instance_with_options, query, MockApi,
+    instantiate, mock_env, mock_info, mock_instance_with_options, query, MockApi,
     MockInstanceOptions, MockQuerier, MockStorage,
 };
 use cosmwasm_vm::{capabilities_from_csv, Instance};
 
 use cw_controllers::AdminResponse;
-use finality::msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
+use finality::msg::{InstantiateMsg, QueryMsg};
 use finality::state::config::Config;
 
 static WASM: &[u8] = include_bytes!("../../../artifacts/finality.wasm");
@@ -37,7 +37,6 @@ fn instantiate_works() {
     let msg = InstantiateMsg {
         admin: mock_api.addr_make(CREATOR),
         bsn_id: "op-stack-l2-11155420".to_string(),
-        is_enabled: false,
         min_pub_rand: 100,
     };
     let info = mock_info(CREATOR, &[]);
@@ -55,90 +54,4 @@ fn instantiate_works() {
     let res: AdminResponse =
         from_json(query(&mut deps, mock_env(), QueryMsg::Admin {}).unwrap()).unwrap();
     assert_eq!(mock_api.addr_make(CREATOR), res.admin.unwrap());
-
-    // Check the contract is disabled on instantiation
-    let enabled: bool =
-        from_json(query(&mut deps, mock_env(), QueryMsg::IsEnabled {}).unwrap()).unwrap();
-    assert!(!enabled);
-}
-
-#[test]
-fn disable_and_reenable_works() {
-    // Setup
-    let mut instance = mock_instance_on_babylon(WASM, &[]);
-    let mock_api = MockApi::default();
-    let msg = InstantiateMsg {
-        admin: mock_api.addr_make(CREATOR),
-        bsn_id: "op-stack-l2-11155420".to_string(),
-        is_enabled: false,
-        min_pub_rand: 100,
-    };
-    let info = mock_info(CREATOR, &[]);
-    let mut res: ContractResult<Response> =
-        instantiate(&mut instance, mock_env(), info, msg.clone());
-    assert!(res.is_ok());
-
-    // Check the contract is disabled on instantiation
-    let mut enabled: bool =
-        from_json(query(&mut instance, mock_env(), QueryMsg::IsEnabled {}).unwrap()).unwrap();
-    assert!(!enabled, "Contract should be disabled on instantiation");
-
-    // Enable the contract
-    let info = mock_info(&mock_api.addr_make(CREATOR), &[]);
-    res = execute(
-        &mut instance,
-        mock_env(),
-        info,
-        ExecuteMsg::SetEnabled { enabled: true },
-    );
-    assert!(res.is_ok());
-    enabled = from_json(query(&mut instance, mock_env(), QueryMsg::IsEnabled {}).unwrap()).unwrap();
-    assert!(enabled, "Enabling works");
-
-    // Disable the contract
-    let info = mock_info(&mock_api.addr_make(CREATOR), &[]);
-    res = execute(
-        &mut instance,
-        mock_env(),
-        info,
-        ExecuteMsg::SetEnabled { enabled: false },
-    );
-    assert!(res.is_ok());
-    enabled = from_json(query(&mut instance, mock_env(), QueryMsg::IsEnabled {}).unwrap()).unwrap();
-    assert!(!enabled, "Disabling works");
-
-    // Re-enable the contract
-    // This simulates a scenario where the contract is disabled for e.g. to fix a bug, and then
-    // subsequently re-enabled after a fix is deployed.
-    let info = mock_info(&mock_api.addr_make(CREATOR), &[]);
-    res = execute(
-        &mut instance,
-        mock_env(),
-        info,
-        ExecuteMsg::SetEnabled { enabled: true },
-    );
-    assert!(res.is_ok());
-    enabled = from_json(query(&mut instance, mock_env(), QueryMsg::IsEnabled {}).unwrap()).unwrap();
-    assert!(enabled, "Re-enabling works");
-}
-
-#[test]
-fn instantiate_enabled() {
-    // Setup
-    let mut instance = mock_instance_on_babylon(WASM, &[]);
-    let mock_api = MockApi::default();
-    let msg = InstantiateMsg {
-        admin: mock_api.addr_make(CREATOR),
-        bsn_id: "op-stack-l2-11155420".to_string(),
-        is_enabled: true,
-        min_pub_rand: 100,
-    };
-    let info = mock_info(CREATOR, &[]);
-    let res: ContractResult<Response> = instantiate(&mut instance, mock_env(), info, msg.clone());
-    assert!(res.is_ok());
-
-    // Check the contract is disabled on instantiation
-    let enabled: bool =
-        from_json(query(&mut instance, mock_env(), QueryMsg::IsEnabled {}).unwrap()).unwrap();
-    assert!(enabled, "Contract should be enabled on instantiation");
 }