+++ CHANGELOG.md
@@ -39,8 +39,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### State and API breaking
 
+* [#65](https://github.com/babylonlabs-io/rollup-bsn-contracts/pull/65) Remove `is_enabled` flag and associated functionality from finality contract
 * [#40](https://github.com/babylonlabs-io/rollup-bsn-contracts/pull/40) Timestamped public randomness commitments.
   Breaks `PubRandCommit` struct that is being used for both storage and queries.
+
+### API breaking
+
+* [#60](https://github.com/babylonlabs-io/rollup-bsn-contracts/pull/60) feat: add signing context
 
 ### State breaking
 
@@ -50,6 +55,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Improvements
 
+* [#62](https://github.com/babylonlabs-io/rollup-bsn-contracts/pull/62) Add consumer ID format validation and admin address validation during contract instantiation and admin updates
 * [#56](https://github.com/babylonlabs-io/rollup-bsn-contracts/pull/56) e2e: migrate e2e from babylon to contract repo
 * [#57](https://github.com/babylonlabs-io/rollup-bsn-contracts/pull/57) Fix block voters response type annotation
 * [#51](https://github.com/babylonlabs-io/rollup-bsn-contracts/pull/51) finality: refactor handle_finality_signature and revise unit tests